requirements.txt
huggingface_hub>=0.20.0
python-dotenv>=1.0.0
Pillow>=10.0.0
flask>=3.0.0
openai>=1.0.0
requests>=2.0.0
